Move search-sort-container into genreGridContainer
