+++ src/App.css
@@ -40,7 +40,7 @@ body {
   width: 100%;
   max-width: 600px;
   align-self: center;
-  margin: 10px auto 20px auto; 
+  margin: 10px auto 10px auto; 
 }
 
 .title-container {
@@ -62,65 +62,10 @@ body {
 }
 
 .page-title {
-  /* flex-grow: 1; */
   font-size: 24px;
   font-weight: bold;
   margin: 10px 0;
   text-align: left;
-}
-
-.search-sort-container {
-  /* width: 100%; */
-  display: flex;
-  justify-content: space-between;
-  /* align-items: center; */
-  gap: 10px; /* Space between search bar, dropdown, and refresh button */
-  margin-bottom: 5px;
-}
-
-.search-bar {
-  width: 85%;
-  max-width: 400px;
-  padding: 10px;
-  /* margin: 20px auto; */
-  border: none;
-  border-radius: 50px;
-  background-color: #282828;
-  color: #ffffff;
-  font-size: 16px;
-  outline: none;
-  transition: background-color 0.3s ease;
-}
-
-.search-bar::placeholder {
-  color: #b3b3b3;
-}
-
-.search-bar:focus {
-  background-color: #3e3e3e;
-}
-
-.sort-dropdown {
-  padding: 10px;
-  border: none;
-  border-radius: 10px;
-  background-color: #282828;
-  color: #ffffff;
-  font-size: 16px;
-  outline: none;
-  appearance: none;
-  -webkit-appearance: none;
-  -moz-appearance: none;
-  transition: background-color 0.3s ease;
-}
-
-.sort-dropdown:focus {
-  background-color: #3e3e3e;
-}
-
-.sort-dropdown option {
-  background-color: #121212;
-  color: #ffffff;
 }
 
 .albums-container {
@@ -133,8 +78,4 @@ body {
 .no-albums {
   font-size: 20px;
   color: #b3b3b3;
-}
-
-@media (max-width: 768px) {
-  
 }
